refactor: reuse canonical path and token helpers in session tools and credentials
